Use JOCL 2.0.5
JOCL 2.0.5 has been the stable release for some time. Including this old release seems to clobber newer versions included with other plugins that use JOCL, due to the priority assigned by the ImageJ updater.

It might be worth considering updating the scijava version in this parent pom as it's quite old now.